move ioreadall and timeafter out of the analysis directory
This is in preparation of making these analyzers separate Go modules.
