+++ ioreadall/io_readall.go
