+++ ioreadall/io_readall_test.go
