+++ ioreadall/testdata/io_read_all_present.go
